moved logic from /auth to /
